requirements.txt
python-telegram-bot==13.15
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
psutil==5.9.7
schedule==1.2.0